Added notification when a new version is available
App now periodically pings a JSON file on encryptr.org and checks if there is a newer version. If there is, it alerts the user and prompts them o download it (sending them to the website).

New versions can be optional, important, or urgent. optional releases are slightly less naggy (no prompt on login).

Fixes #116