Avoid char encoding issues and XAS
In order to avoid both XAS and passwords with chars that wouldn't be
copied to the clipboard, instead of taking the text from the value
field, add the key to the field in a data-attr, then use it to fetch the
value from the modeland load that onto the clipboard.

Then (for the XAS), make sure to use the {{! blah }} syntax of DoT to
encode the value before displaying it.